move UdpConnection struct and Socket enum to net_utils
